libzeth: move some general code out of mpc dir
