+++ libzeth/core/chacha_rng.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: LGPL-3.0+
 
-#include "libzeth/snarks/groth16/mpc/chacha_rng.hpp"
+#include "libzeth/core/chacha_rng.hpp"
 
 #include <algorithm>
 #include <cstring>
