+++ libzeth/snarks/groth16/mpc/mpc_utils.tcc
@@ -5,10 +5,10 @@
 #ifndef __ZETH_SNARKS_GROTH16_MPC_UTILS_TCC__
 #define __ZETH_SNARKS_GROTH16_MPC_UTILS_TCC__
 
+#include "libzeth/core/evaluator_from_lagrange.hpp"
+#include "libzeth/core/multi_exp.hpp"
 #include "libzeth/core/utils.hpp"
-#include "libzeth/snarks/groth16/mpc/evaluator_from_lagrange.hpp"
 #include "libzeth/snarks/groth16/mpc/mpc_utils.hpp"
-#include "libzeth/snarks/groth16/mpc/multi_exp.hpp"
 #include "libzeth/snarks/groth16/mpc/phase2.hpp"
 
 #include <algorithm>
