+++ libzeth/test/mpc_test.cpp
@@ -4,12 +4,12 @@
 
 #include "libzeth/circuits/circuit_types.hpp"
 #include "libzeth/circuits/sha256/sha256_ethereum.hpp"
+#include "libzeth/core/chacha_rng.hpp"
+#include "libzeth/core/evaluator_from_lagrange.hpp"
+#include "libzeth/core/multi_exp.hpp"
 #include "libzeth/core/utils.hpp"
 #include "libzeth/snarks/groth16/groth16_core.hpp"
-#include "libzeth/snarks/groth16/mpc/chacha_rng.hpp"
-#include "libzeth/snarks/groth16/mpc/evaluator_from_lagrange.hpp"
 #include "libzeth/snarks/groth16/mpc/mpc_utils.hpp"
-#include "libzeth/snarks/groth16/mpc/multi_exp.hpp"
 #include "libzeth/snarks/groth16/mpc/phase2.hpp"
 #include "libzeth/snarks/groth16/mpc/powersoftau_utils.hpp"
 #include "libzeth/test/simple_test.hpp"
