+++ libzeth/test/powersoftau_test.cpp
@@ -2,8 +2,8 @@
 //
 // SPDX-License-Identifier: LGPL-3.0+
 
+#include "libzeth/core/evaluator_from_lagrange.hpp"
 #include "libzeth/core/utils.hpp"
-#include "libzeth/snarks/groth16/mpc/evaluator_from_lagrange.hpp"
 #include "libzeth/snarks/groth16/mpc/powersoftau_utils.hpp"
 
 #include <boost/filesystem.hpp>
